feat: clean up aptatrans FCS extraction
